Rename and refactor RdsCompFluentGrid to RdsCompGrid
The `RdsCompFluentGrid` component has been renamed to `RdsCompGrid` across all files, reflecting a rebranding effort. Import statements and TypeScript interfaces have been updated accordingly.

Significant enhancements include the introduction of new features such as row and column reordering via drag-and-drop, advanced filtering, sorting, and inline editing capabilities. The SCSS file has been cleaned up and modernized with new CSS variables for light and dark themes, following a BEM-like convention for improved maintainability.

Documentation has been updated to reflect the new component name and features, with enhanced Storybook examples demonstrating various functionalities. The overall structure and responsiveness of the component have been improved for a better user experience.
