+++ raaghu-components/index.ts
@@ -45,4 +45,3 @@ export { default as RdsCompAiGradientTextWithIcon } from './rds-comp-ai-gradient
 export { default as RdsCompAiMessageBox } from './rds-comp-ai-message-box/rds-comp-ai-message-box';
 export { default as RdsCompAiChatHeader } from './rds-comp-ai-chat-header/rds-comp-ai-chat-header';
 export { default as RdsCompDatePicker } from './rds-comp-date-and-time-picker/rds-comp-date-and-time-picker';
-export { default as RdsCompFluentGrid } from './rds-comp-fluent-grid/rds-comp-fluent-grid';
